# src/Innovark.Weather.UI/Innovark.Weather.Web/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Innovark.Weather.Web | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/weather/history": {
      "post": {
        "tags": [
          "Weather"
        ],
        "summary": "Get the 10 hourly records for a date and hour in UTC+7 from the weather API",
        "operationId": "PostWeatherHistory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherHistoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "WeatherHistoryRecord": {
        "required": [
          "time",
          "temperatureC",
          "temperatureF",
          "relativeHumidity"
        ],
        "type": "object",
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "number",
            "format": "double"
          },
          "temperatureF": {
            "type": "number",
            "format": "double"
          },
          "relativeHumidity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "WeatherHistoryRequest": {
        "required": [
          "date",
          "hour"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "hour": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "WeatherHistoryResponse": {
        "required": [
          "requestedTime",
          "records"
        ],
        "type": "object",
        "properties": {
          "requestedTime": {
            "type": "string",
            "format": "date-time"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeatherHistoryRecord"
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Weather"
    }
  ]
}